aligned chips
